[Fix] Top results permanently filtered.
Condition: When podcast is favorited, the podcast is filtered from the
top podcasts that fill the same category.
Bug: After removing podcast from favorites, it is still filtered from
the top podcasts.
Reason: badcachee.aggregate_cat incorrectly assigned reference to
top100cat[genre] to temporary variable.
Fix: Used Array.slice() to ensure copy of array is passed to temporary
variable, and not the reference.